+++ CHANGELOG.md
@@ -45,6 +45,7 @@ straightforward as possible.
 - cargo room unblock [#504]
 - handle FileReader on stopped db [#505]
 - correct Chat.TaskSupervisor usage [#512]
+- upload pruner error [#517]
 
 ## 2023-08-11_279815b___2023-08-11_24d6bac